Changed int to decimal in WireParameters

+++ Urania/Urania.Core/WireParameters.cs
@@ -1,8 +1,8 @@
 ﻿namespace Urania.Core {
     public class WireParameters {
-        public int Wd { get; set; }
-        public int Id { get; set; }
-        public int Od { get; set; }
-        public int Ar { get; set; }
+        public decimal Wd { get; set; }
+        public decimal Id { get; set; }
+        public decimal Od { get; set; }
+        public decimal Ar { get; set; }
     }
 }
